MagicCombo: Timers are fine with only 8 bits
There is no use for timer-outs longer than the 256 cycles a 8-bit timer
allows. As such, confine the MagicCombo timers to 8-bits too, saving us
a few bytes.

+++ lib/Akela-MagicCombo/src/Akela/MagicCombo.cpp
@@ -21,8 +21,8 @@
 namespace Akela {
 
   const MagicCombo::dictionary_t *MagicCombo::dictionary;
-  uint16_t MagicCombo::timeOut;
-  uint16_t MagicCombo::timer;
+  uint8_t MagicCombo::timeOut;
+  uint8_t MagicCombo::timer;
 
   MagicCombo::MagicCombo (void) {
   }
